Refactor RuleAlteredJobProgressDetector for common usage
