+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/api/job/progress/PipelineJobItemProgress.java
@@ -17,8 +17,17 @@
 
 package org.apache.shardingsphere.data.pipeline.api.job.progress;
 
+import org.apache.shardingsphere.data.pipeline.api.job.JobStatus;
+
 /**
  * Pipeline job item progress.
  */
 public interface PipelineJobItemProgress {
+    
+    /**
+     * Get job status.
+     *
+     * @return job status
+     */
+    JobStatus getStatus();
 }
